Added stumble on close call lane switch
The way this works is by utilising a ray cast. Every time the player switches lane, we do a ray cast in front of the player in the lane they're switching away from (the distance of it is tied to world speed) and if it hits something we invoke a stumble lane change. Otherwise, it is considered a smooth lane change.